backend/requirements.txt
bcrypt==3.1.6
cffi==1.12.3
Django==2.2
django-webpack-loader==0.6.0
ecdsa==0.13.2
future==0.17.1
psycopg2==2.8.1
pyasn1==0.4.5
pycparser==2.19
python-jose==3.0.1
pytz==2018.9
rsa==4.0
six==1.12.0
sqlparse==0.3.0
